+++ trend-collector/trendcollector/libs/infrastractures/__init__.py
@@ -1,4 +1,3 @@
 from .conf import *
-from .schemas import Base
 from .trend_repository import TrendRepository
-from .twitter_account_repository import *
+from .repositories import *
