+++ trend-collector/trendcollector/libs/infrastractures/repositories/__init__.py
@@ -1,0 +1,1 @@
+from .twitter_account import TwitterAccountRepository
