+++ trend-collector/trendcollector/libs/infrastractures/repositories/schemas.py
@@ -1,16 +1,43 @@
+import functools
+from http import HTTPStatus
+
 from sqlalchemy import DATETIME, Column, String
 from sqlalchemy.dialects.mysql import INTEGER as Integer
+from sqlalchemy.exc import InvalidRequestError, OperationalError as SQLAlchemyOperationalError
 from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.orm.exc import DetachedInstanceError
 from sqlalchemy.sql.expression import text
 from sqlalchemy.sql.functions import current_timestamp
 
-TWITTER_ACCOUNTS = "twitter_accounts"
+from ...services.accessor import (AttributesException, DetachedInstance, InvalidRequestException,
+                                  OperationalException, RuntimeException)
+
+
+# コールバック関数の引数(*args, **kwargs)をCallableで表現することは不可能なので型ヒントは書かない
+def handle_exception(func):
+    @functools.wraps(func)
+    def _handler_wrapper(*args, **kwargs):
+        try:
+            return func(*args, **kwargs)
+        except DetachedInstanceError as e:
+            raise DetachedInstance(HTTPStatus.INTERNAL_SERVER_ERROR, "", list(e.args))
+        except SQLAlchemyOperationalError as e:
+            raise OperationalException(HTTPStatus.INTERNAL_SERVER_ERROR, "", list(e.args))
+        except InvalidRequestError as e:
+            raise InvalidRequestException(HTTPStatus.INTERNAL_SERVER_ERROR, "", list(e.args))
+        except AttributeError as e:
+            raise AttributesException(HTTPStatus.BAD_REQUEST, "", list(e.args))
+        except RuntimeError as e:
+            raise RuntimeException(HTTPStatus.INTERNAL_SERVER_ERROR, "", list(e.args))
+
+    return _handler_wrapper
+
 
 Base = declarative_base()
 
 
 class TwitterAccountTable(Base):
-    __tablename__ = TWITTER_ACCOUNTS
+    __tablename__ = "twitter_accounts"
 
     id = Column(Integer, primary_key=True)
     account_id = Column(Integer)
